+++ mml/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2024-10-27
+
 ### Added
 
 - Added `tokio` and `async-std` cargo features (they are both mutually exclusive).